requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-docx==1.1.0
qrcode[pil]==7.4.2
Pillow==10.1.0
